git-svn-id: https://priceme.unfuddle.com/svn/priceme_minitools@380 6c2e4e5e-b397-40cf-afe5-66bc484122c4

+++ Tian/HotterWinds_1.4/HotterWinds/Modules/Products/ProductTopDisplay.ascx.cs
@@ -53,7 +53,7 @@ namespace HotterWinds.Modules.Products
 
         public int InStockStatus = 0;
 
-        public Tuple<int, int> ReviewsRatings = null;
+        public Tuple<double, int> ReviewsRatings = null;
 
         protected void Page_Load(object sender, EventArgs e)
         {
@@ -70,7 +70,7 @@ namespace HotterWinds.Modules.Products
             ReviewsRatings = DBQuery.ProductQuery.GetProductRating(Product.ProductID);
             if (ReviewsRatings == null)
             {
-                ReviewsRatings = new Tuple<int, int>(0, 0);
+                ReviewsRatings = new Tuple<double, int>(0, 0);
             }
 
             //ProductRatingControl1.productId = Product.ProductID;
